chore: Initialize Application with Svelte Kit
We simply ran `npm create svelte@latest .` and
installed the packages.

We also reconciled Svelte Kit's default files
with our original files (this only applies to
the `README` file and the `.gitignore` file).

Oh. We also udpated the `prettier` config to
fit our tastes (after which we ran the formatter).