chore: Add Test Cases for the Authentication Features
This was important to do to show devs what features we
support and to give the SuperTokens team a template of
things that could be helpful to test.

However, thanks to these Test Cases, we also found bugs
in our application that we were able to fix. Maybe this
is where something like Cypress or Playwrighth could be
useful.

There are a few small test cases that still need to be
verified. These are marked with empty checkboxes.