chore: Copy automated tests over from Remix Version of the Project
Pretty much all of the tests are passing. Only one of the tests is
failing because Svelte does not automatically convert `boolean`s
passed to `aria-invalid` to `"true"`/`"false"`. We need to figure
out if we want to raise this to Svelte's attention as a GH issue,
or if we just want to handle the problem ourselves. We'd rather
not change the tests, so we'd have to add some additional JS logic
in ourselves if we didn't open a new issue.

Note that invalid `aria-invalid` attribute values are treated as
`"true"`. However, we'd rather have Svelte generate a valid value
for us than rely on this logic. Being explicit is often best.

In any case, until this issue is fixed, you'll have to run the tests
in a more manual manner. You won't be able to do "Run All" quite yet.